#0: split data movement ops into hpp and cpp
